Add files via upload
AntiColor files contain, as usual, a ball, and a brick. In these files, the ball will only stop falling if the pixel below it is NOT the color requested.
In the other file, Rolling, it is the same case, but the ball moves around on the brick, going from side to side, until it finds a possible hole.